Convert request body from HttpEntity to InputStream
The ring spec [1] states that the request body should be of type
InputStream.  However, clj-http.client/post and similar wrap their
request bodies in an HttpEntity subclass such as StringEntity, meaning
that clj-http-fake doesn't work for POSTs and PUTs and other requests
which include a body.

Fixes #16.

Co-authored by @rnewstead1.

[1] https://github.com/ring-clojure/ring/blob/master/SPEC